spics: Add wrapper to work around spidev
spidev doesn't necessarily allow keeping the CS line active after an operation
completes. Some devices interpret this as the end of a transaction. The problem
is that spidev has a limited buffer size (default: 4096 bytes) so it may become
impossible to successfully complete a long logical transaction.

Add aliases for CS pins. They normally fail to be registered correctly.